dreamvision-dashboard/backend/requirements.txt
fastapi
uvicorn
python-multipart
passlib[bcrypt]
pyjwt
sqlalchemy
websockets
opencv-python-headless
numpy
requests
fpdf2
